refactor: 리뷰반영 - 패키지 분류 변경
