+++ src/test/java/carracing/InputInfoTest.java
@@ -1,6 +1,6 @@
 package carracing;
 
-import carracing.domain.InputInfo;
+import carracing.view.InputInfo;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
